Java version downgrade to 17 LTS; Aquality version updated; project structure corrections; style corrections
