--- src/test/java/framework/LocatorConstants.java
@@ -1,6 +1,0 @@
-package framework;
-
-public class LocatorConstants {
-    public static final String PRECISE_TEXT_XPATH = "//*[text()='%s']";
-    public static final String PARTICULAR_TEXT_XPATH = "//*[contains(text(),'%s')]";
-}
