--- src/test/java/framework/models/DataTableData.java
@@ -1,8 +1,0 @@
-package framework.models;
-
-import lombok.Getter;
-
-@Getter
-public class DataTableData {
-    private Double expectedDueSum;
-}
