--- src/test/java/framework/models/Env.java
@@ -1,8 +1,0 @@
-package framework.models;
-
-import lombok.Getter;
-
-@Getter
-public class Env {
-    private String env;
-}
