--- src/test/java/framework/models/EnvData.java
@@ -1,15 +1,0 @@
-package framework.models;
-
-import lombok.Getter;
-
-@Getter
-public class EnvData {
-    private String protocol;
-    private String domain;
-    private int wait;
-    private String host;
-
-    public String getHost() {
-        return protocol + "://" + domain;
-    }
-}
