--- src/test/java/framework/models/FileData.java
@@ -1,9 +1,0 @@
-package framework.models;
-
-import lombok.Getter;
-
-@Getter
-public class FileData {
-    private String uploadFile;
-    private String downloadFile;
-}
