--- src/test/java/framework/models/UserData.java
@@ -1,9 +1,0 @@
-package framework.models;
-
-import lombok.Getter;
-
-@Getter
-public class UserData {
-    private String username;
-    private String password;
-}
