--- src/test/java/framework/pages/DynamicControlsPage.java
@@ -1,18 +1,0 @@
-package framework.pages;
-
-import aquality.selenium.browser.AqualityServices;
-import aquality.selenium.elements.interfaces.IButton;
-import aquality.selenium.elements.interfaces.IElementFactory;
-import aquality.selenium.elements.interfaces.ITextBox;
-import aquality.selenium.forms.Form;
-import framework.LocatorConstants;
-import org.openqa.selenium.By;
-
-public class DynamicControlsPage extends Form {
-
-    // todo: implement class
-    public DynamicControlsPage() {
-        // todo: implement
-        super(By.xpath(""), "");
-    }
-}
