--- src/test/java/framework/tests/BaseTest.java
@@ -1,27 +1,0 @@
-package framework.tests;
-
-import aquality.selenium.browser.AqualityServices;
-import aquality.selenium.browser.Browser;
-import framework.pages.MainPage;
-import framework.utils.SettingsTestData;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
-
-public class BaseTest {
-    protected MainPage mainPage = new MainPage();
-    protected Browser browser;
-
-
-
-    @BeforeMethod
-    public void setup() {
-        browser = AqualityServices.getBrowser();
-        browser.maximize();
-        browser.goTo(SettingsTestData.getEnvData().getHost());
-    }
-
-    @AfterMethod
-    public void teardown() {
-        browser.quit();
-    }
-}
