--- src/test/java/framework/tests/BasicAuthTest.java
@@ -1,22 +1,0 @@
-package framework.tests;
-
-import framework.pages.BasicAuthPage;
-import framework.pages.MainPageNavigation;
-import framework.utils.SettingsTestData;
-import org.testng.Assert;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
-
-public class BasicAuthTest extends BaseTest {
-    private BasicAuthPage basicAuthPage = new BasicAuthPage();
-
-    @BeforeMethod
-    public void auth() {
-        // todo: add basic auth
-    }
-
-    @Test
-    public void basicAuthTest() {
-        // todo: add test
-    }
-}
