--- src/test/java/framework/tests/DynamicControlsTest.java
@@ -1,20 +1,0 @@
-package framework.tests;
-
-import framework.pages.DynamicControlsPage;
-import framework.pages.MainPageNavigation;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-import java.util.UUID;
-
-public class DynamicControlsTest extends BaseTest {
-    private DynamicControlsPage dynamicControlsPage = new DynamicControlsPage();
-    private final String RANDOM_TEXT = UUID.randomUUID().toString();
-
-    @Test
-    public void dynamicControlsTest() {
-        // todo: add test
-    }
-
-
-}
