+++ src/test/java/models/DataTableData.java
@@ -1,0 +1,8 @@
+package models;
+
+import lombok.Getter;
+
+@Getter
+public class DataTableData {
+    private Double expectedDueSum;
+}
