+++ src/test/java/models/Env.java
@@ -1,0 +1,8 @@
+package models;
+
+import lombok.Getter;
+
+@Getter
+public class Env {
+    private String env;
+}
