+++ src/test/java/models/EnvData.java
@@ -1,0 +1,15 @@
+package models;
+
+import lombok.Getter;
+
+@Getter
+public class EnvData {
+    private String protocol;
+    private String domain;
+    private int wait;
+    private String host;
+
+    public String getHost() {
+        return protocol + "://" + domain;
+    }
+}
