+++ src/test/java/models/FileData.java
@@ -1,0 +1,9 @@
+package models;
+
+import lombok.Getter;
+
+@Getter
+public class FileData {
+    private String uploadFile;
+    private String downloadFile;
+}
