+++ src/test/java/models/UserData.java
@@ -1,0 +1,9 @@
+package models;
+
+import lombok.Getter;
+
+@Getter
+public class UserData {
+    private String username;
+    private String password;
+}
