+++ src/test/java/pages/DynamicControlsPage.java
@@ -1,0 +1,13 @@
+package pages;
+
+import aquality.selenium.forms.Form;
+import org.openqa.selenium.By;
+
+public class DynamicControlsPage extends Form {
+
+    // todo: implement class
+    public DynamicControlsPage() {
+        // todo: implement
+        super(By.xpath(""), "");
+    }
+}
