+++ src/test/java/tests/BaseTest.java
@@ -1,0 +1,26 @@
+package tests;
+
+import aquality.selenium.browser.AqualityServices;
+import pages.MainPage;
+import utils.SettingsTestData;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+import static aquality.selenium.browser.AqualityServices.getBrowser;
+
+public abstract class BaseTest {
+    protected final MainPage mainPage = new MainPage();
+
+    @BeforeMethod
+    public void setup() {
+        getBrowser().maximize();
+        getBrowser().goTo(SettingsTestData.getEnvData().getHost());
+    }
+
+    @AfterMethod
+    public void teardown() {
+        if (AqualityServices.isBrowserStarted()) {
+            getBrowser().quit();
+        }
+    }
+}
