+++ src/test/java/tests/BasicAuthTest.java
@@ -1,0 +1,19 @@
+package tests;
+
+import pages.BasicAuthPage;
+import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Test;
+
+public class BasicAuthTest extends BaseTest {
+    private final BasicAuthPage basicAuthPage = new BasicAuthPage();
+
+    @BeforeMethod
+    public void auth() {
+        // todo: add basic auth
+    }
+
+    @Test
+    public void basicAuthTest() {
+        // todo: add test
+    }
+}
