+++ src/test/java/tests/DynamicControlsTest.java
@@ -1,0 +1,16 @@
+package tests;
+
+import pages.DynamicControlsPage;
+import org.testng.annotations.Test;
+
+import java.util.UUID;
+
+public class DynamicControlsTest extends BaseTest {
+    private final DynamicControlsPage dynamicControlsPage = new DynamicControlsPage();
+    private final String RANDOM_TEXT = UUID.randomUUID().toString();
+
+    @Test
+    public void dynamicControlsTest() {
+        // todo: add test
+    }
+}
