--- src/test/resources/dataTableData.json
@@ -1,3 +1,0 @@
-{
-  "expectedDueSum": 251.0
-}
