--- src/test/resources/file.json
@@ -1,1 +1,0 @@
-{"key": "value"}
