--- src/test/resources/fileData.json
@@ -1,4 +1,0 @@
-{
-  "uploadFile": "file.json",
-  "downloadFile": "Test.txt"
-}
