+++ src/test/resources/settings.json
