+++ src/test/resources/testdata/dataTableData.json
@@ -1,0 +1,3 @@
+{
+  "expectedDueSum": 251.0
+}
