+++ src/test/resources/testdata/file.json
@@ -1,0 +1,3 @@
+{
+  "key": "value"
+}
