+++ src/test/resources/testdata/fileData.json
@@ -1,0 +1,4 @@
+{
+  "uploadFile": "file.json",
+  "downloadFile": "Test.txt"
+}
