+++ src/test/resources/testdata/userData.json
@@ -1,0 +1,4 @@
+{
+  "username": "admin",
+  "password": "admin"
+}
