--- src/test/resources/userData.json
@@ -1,4 +1,0 @@
-{
-  "username": "admin",
-  "password": "admin"
-}
